cleanup
Former-commit-id: aef44f1067eedd7fe02824c0ebc7bd68aded2a21

+++ user/chenyk/cemdutil1.c
@@ -67,8 +67,8 @@ typedef struct {
 #define EMD_IO_H
 
 typedef struct {
-    double threshold;
-    double tolerance;
+    float threshold;
+    float tolerance;
 } stop_t;
 /*^*/
